V.7

+++ src/task3.cpp
@@ -5,7 +5,7 @@
 #include "task2.h"
 
 unsigned long long sumPrime(unsigned int hbound){
-    unsigned long sum = 0;
+    unsigned long long sum = 0;
     for (unsigned int i = 2; i < hbound; ++i)
         if (checkPrime(i)) {
             sum += i;
